refactor: moved fixtures in their own module
