+++ tests/units/__init__.py
@@ -2,3 +2,9 @@
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
 """Unit tests for ANTA."""
+
+import pytest
+
+# Enable nice assert messages for tests.units.anta_tests unit tests
+# https://docs.pytest.org/en/stable/how-to/writing_plugins.html#assertion-rewriting
+pytest.register_assert_rewrite("tests.units.anta_tests")
